+++ src/main/java/com/zendesk/maxwell/producer/AbstractProducer.java
@@ -17,7 +17,8 @@ public abstract class AbstractProducer {
 	protected final Meter succeededMessageMeter;
 	protected final Counter failedMessageCount;
 	protected final Meter failedMessageMeter;
-	protected final Timer metricsTimer;
+	protected final Timer messagePublishTimer;
+	protected final Timer messageLatencyTimer;
 
 	public AbstractProducer(MaxwellContext context) {
 		this.context = context;
@@ -30,7 +31,8 @@ public abstract class AbstractProducer {
 		this.succeededMessageMeter = metricRegistry.meter(metrics.metricName("messages", "succeeded", "meter"));
 		this.failedMessageCount = metricRegistry.counter(metrics.metricName("messages", "failed"));
 		this.failedMessageMeter = metricRegistry.meter(metrics.metricName("messages", "failed", "meter"));
-		this.metricsTimer = metrics.getRegistry().timer(metrics.metricName("message", "publish", "time"));
+		this.messagePublishTimer = metrics.getRegistry().timer(metrics.metricName("message", "publish", "time"));
+		this.messageLatencyTimer = metrics.getRegistry().timer(metrics.metricName("message", "publish", "age"));
 	}
 
 	abstract public void push(RowMap r) throws Exception;