+++ src/main/java/com/zendesk/maxwell/row/RowMap.java
@@ -30,7 +30,7 @@ public class RowMap implements Serializable {
 
 	static final Logger LOGGER = LoggerFactory.getLogger(RowMap.class);
 
-	private final String rowQuery;
+	private String rowQuery;
 	private final String rowType;
 	private final String database;
 	private final String table;
@@ -46,6 +46,7 @@ public class RowMap implements Serializable {
 	private boolean txCommit;
 	private Long serverId;
 	private Long threadId;
+	private Long schemaId;
 
 	private final LinkedHashMap<String, Object> data;
 	private final LinkedHashMap<String, Object> oldData;
@@ -294,7 +295,6 @@ public class RowMap implements Serializable {
 		if ( outputConfig.includesBinlogPosition )
 			g.writeStringField(FieldNames.POSITION, binlogPosition.getFile() + ":" + binlogPosition.getOffset());
 
-
 		if ( outputConfig.includesGtidPosition)
 			g.writeStringField(FieldNames.GTID, binlogPosition.getGtid());
 
@@ -304,6 +304,10 @@ public class RowMap implements Serializable {
 
 		if ( outputConfig.includesThreadId && this.threadId != null ) {
 			g.writeNumberField(FieldNames.THREAD_ID, this.threadId);
+		}
+
+		if ( outputConfig.includesSchemaId && this.schemaId != null ) {
+			 g.writeNumberField(FieldNames.SCHEMA_ID, this.schemaId);
 		}
 
 		for ( Map.Entry<String, Object> entry : this.extraAttributes.entrySet() ) {
@@ -457,6 +461,14 @@ public class RowMap implements Serializable {
 		this.threadId = threadId;
 	}
 
+	public Long getSchemaId() {
+		return schemaId;
+	}
+
+	public void setSchemaId(Long schemaId) {
+		this.schemaId = schemaId;
+	}
+
 	public String getDatabase() {
 		return database;
 	}
@@ -479,6 +491,10 @@ public class RowMap implements Serializable {
 
 	public String getRowQuery() {
 		return rowQuery;
+	}
+
+	public void setRowQuery(String query) {
+		this.rowQuery = query;
 	}
 
 	public String getRowType() {