+++ src/main/java/com/zendesk/maxwell/schema/Table.java
@@ -1,6 +1,7 @@
 package com.zendesk.maxwell.schema;
 
 import java.util.*;
+import java.util.stream.Collectors;
 
 import com.zendesk.maxwell.schema.ddl.InvalidSchemaError;
 import com.zendesk.maxwell.schema.ddl.ColumnPosition;
@@ -25,21 +26,28 @@ public class Table {
 	private List<String> pkColumnNames;
 	private List<String> normalizedPKColumnNames;
 
-	private HashMap<String, Integer> columnOffsetMap;
 	@JsonIgnore
 	public int pkIndex;
 
 	public Table() { }
 	public Table(String database, String name, String charset, List<ColumnDef> list, List<String> pks) {
-		this.database = database;
-		this.name = name;
+		this.database = database.intern();
+		this.name = name.intern();
 		this.charset = charset;
+		if ( this.charset != null )
+			this.charset = this.charset.intern();
+
 		this.setColumnList(list);
 
 		if ( pks == null )
 			pks = new ArrayList<String>();
 
 		this.setPKList(pks);
+	}
+
+	@JsonProperty("table")
+	public void setTable(String name) {
+		this.name = name.intern();
 	}
 
 	@JsonProperty("columns")
@@ -243,11 +251,17 @@ public class Table {
 	}
 
 	public void setDatabase(String database) {
-		this.database = database;
+		this.database = database.intern();
 	}
 
 	public String getCharset() {
 		return charset;
+	}
+
+	public void setCharset(String charset) {
+		this.charset = charset;
+		if ( this.charset != null )
+			this.charset = charset.intern();
 	}
 
 	@JsonProperty("primary-key")
@@ -265,7 +279,7 @@ public class Table {
 
 	@JsonProperty("primary-key")
 	public synchronized void setPKList(List<String> pkColumnNames) {
-		this.pkColumnNames = pkColumnNames;
+		this.pkColumnNames = pkColumnNames.stream().map((n) -> n.intern()).collect(Collectors.toList());
 		this.normalizedPKColumnNames = null;
 	}