+++ src/main/java/com/zendesk/maxwell/schema/columndef/FloatColumnDef.java
@@ -2,7 +2,7 @@ package com.zendesk.maxwell.schema.columndef;
 
 public class FloatColumnDef extends ColumnDef {
 	public FloatColumnDef() { }
-	public FloatColumnDef(String name, String type, int pos) {
+	public FloatColumnDef(String name, String type, short pos) {
 		super(name, type, pos);
 	}