+++ src/main/java/com/zendesk/maxwell/schema/columndef/JsonColumnDef.java
@@ -8,7 +8,7 @@ import java.io.IOException;
 import static com.github.shyiko.mysql.binlog.event.deserialization.ColumnType.*;
 
 public class JsonColumnDef extends ColumnDef {
-	public JsonColumnDef(String name, String type, int pos) {
+	public JsonColumnDef(String name, String type, short pos) {
 		super(name, type, pos);
 	}