+++ src/main/java/com/zendesk/maxwell/schema/columndef/SetColumnDef.java
@@ -6,7 +6,7 @@ import java.util.Arrays;
 import org.apache.commons.lang3.StringUtils;
 
 public class SetColumnDef extends EnumeratedColumnDef {
-	public SetColumnDef(String name, String type, int pos, String[] enumValues) {
+	public SetColumnDef(String name, String type, short pos, String[] enumValues) {
 		super(name, type, pos, enumValues);
 	}