+++ src/main/java/com/zendesk/maxwell/schema/columndef/YearColumnDef.java
@@ -4,7 +4,7 @@ import java.sql.Date;
 import java.util.Calendar;
 
 public class YearColumnDef extends ColumnDef {
-	public YearColumnDef(String name, String type, int pos) {
+	public YearColumnDef(String name, String type, short pos) {
 		super(name, type, pos);
 	}