+++ src/test/java/com/zendesk/maxwell/schema/SchemaCaptureTest.java
@@ -16,6 +16,7 @@ import com.zendesk.maxwell.MaxwellTestWithIsolatedServer;
 import org.apache.commons.lang3.StringUtils;
 import org.junit.After;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
 
 import static org.hamcrest.CoreMatchers.*;
@@ -180,5 +181,23 @@ public class SchemaCaptureTest extends MaxwellTestWithIsolatedServer {
 		assertEquals("\\,", result[5]);
 		assertEquals(",'", result[6]);
 		assertEquals("b", result[7]);
+	}
+
+	private long getUsedMem() {
+		Runtime r = Runtime.getRuntime();
+		return r.totalMemory() - r.freeMemory();
+	}
+
+	@Ignore
+	@Test
+	public void testHugeCaptureMemUsage() throws Exception {
+		Runtime.getRuntime().gc();
+		System.out.println("usage before: " + getUsedMem());
+		generateHugeSchema();
+		Schema s = capturer.capture();
+		Runtime.getRuntime().gc();
+		Runtime.getRuntime().gc();
+		System.out.println("usage after: " + getUsedMem());
+		System.out.println(s.getCharset());
 	}
 }