+++ src/test/java/com/zendesk/maxwell/schema/ddl/DDLIntegrationTest.java
@@ -402,6 +402,19 @@ public class DDLIntegrationTest extends MaxwellTestWithIsolatedServer {
 	}
 
 	@Test
+	public void testDatabaseAlterMySqlTableCharset() throws Exception {
+		testIntegration("ALTER TABLE mysql.columns_priv " +
+				"MODIFY Host char(60) NOT NULL default '', " +
+				"MODIFY Db char(64) NOT NULL default '', " +
+				"MODIFY User char(16) NOT NULL default '', " +
+				"MODIFY Table_name char(64) NOT NULL default '', " +
+				"MODIFY Column_name char(64) NOT NULL default '', " +
+				"ENGINE=MyISAM, " +
+				"CONVERT TO CHARACTER SET utf8 COLLATE utf8_bin, " +
+				"COMMENT='Column privileges'");
+	}
+
+	@Test
 	@Category(Mysql57Tests.class)
 	public void testGeneratedColumns() throws Exception {
 		requireMinimumVersion(server.VERSION_5_7);
@@ -419,11 +432,26 @@ public class DDLIntegrationTest extends MaxwellTestWithIsolatedServer {
 		assertEquals(1, rows.size());
 		assertTrue(rows.get(0).toJSON(ddlOutputConfig()).contains("\"type\":\"table-create\",\"database\":\"mysql\",\"table\":\"TestTableCreate1\""));
 	}
+	
+	@Test
+	public void testNonLatinTableCreate() throws Exception {
+		String[] sql = {"create table 測試表格 ( 測試欄位一 int, 測試欄位二 text )"};
+		List<RowMap> rows = getRowsForDDLTransaction(sql, null);
+		assertEquals(1, rows.size());
+		assertTrue(rows.get(0).toJSON(ddlOutputConfig()).contains("\"type\":\"table-create\",\"database\":\"mysql\",\"table\":\"測試表格\""));
+	}
 
 	@Test
 	public void testTableCreateFilter() throws Exception {
 		String[] sql = {"create table TestTableCreate2 ( account_id int, text_field text )"};
 		List<RowMap> rows = getRowsForDDLTransaction(sql, excludeTable("TestTableCreate2"));
+		assertEquals(0, rows.size());
+	}
+	
+	@Test
+	public void testNonLatinTableCreateFilter() throws Exception {
+		String[] sql = {"create table 測試表格二 ( 測試欄位一 int, 測試欄位二 text )"};
+		List<RowMap> rows = getRowsForDDLTransaction(sql, excludeTable("測試表格二"));
 		assertEquals(0, rows.size());
 	}
 
@@ -437,6 +465,18 @@ public class DDLIntegrationTest extends MaxwellTestWithIsolatedServer {
 		assertEquals(1, rows.size());
 		assertTrue(rows.get(0).toJSON(ddlOutputConfig()).contains("\"type\":\"table-create\",\"database\":\"mysql\",\"table\":\"TestTableCreate3\""));
 	}
+	
+	@Test
+	public void testNonLatinTableRenameFilter() throws Exception {
+		String[] sql = {
+			"create table 測試表格三 ( 測試欄位一 int, 測試欄位二 text )",
+			"rename table 測試表格三 to 測試表格四"
+		};
+		List<RowMap> rows = getRowsForDDLTransaction(sql, excludeTable("測試表格四"));
+		assertEquals(1, rows.size());
+		assertTrue(rows.get(0).toJSON(ddlOutputConfig()).contains("\"type\":\"table-create\",\"database\":\"mysql\",\"table\":\"測試表格三\""));
+	}
+	
 
 	@Test
 	public void testDatabaseCreate() throws Exception {
@@ -449,6 +489,18 @@ public class DDLIntegrationTest extends MaxwellTestWithIsolatedServer {
 		assertTrue(rows.get(0).toJSON(ddlOutputConfig()).contains("\"type\":\"database-create\",\"database\":\"TestDatabaseCreate1\""));
 		assertTrue(rows.get(1).toJSON(ddlOutputConfig()).contains("\"type\":\"database-alter\",\"database\":\"TestDatabaseCreate1\""));
 	}
+	
+	@Test
+	public void testNonLatinDatabaseCreate() throws Exception {
+		String[] sql = {
+			"create database 測試資料庫一",
+			"alter database 測試資料庫一 character set latin2"
+		};
+		List<RowMap> rows = getRowsForDDLTransaction(sql, null);
+		assertEquals(2, rows.size());
+		assertTrue(rows.get(0).toJSON(ddlOutputConfig()).contains("\"type\":\"database-create\",\"database\":\"測試資料庫一\""));
+		assertTrue(rows.get(1).toJSON(ddlOutputConfig()).contains("\"type\":\"database-alter\",\"database\":\"測試資料庫一\""));
+	}
 
 	@Test
 	public void testDatabaseFilter() throws Exception {
@@ -456,6 +508,14 @@ public class DDLIntegrationTest extends MaxwellTestWithIsolatedServer {
 		List<RowMap> rows = getRowsForDDLTransaction(sql, excludeDb("TestDatabaseCreate2"));
 		assertEquals(0, rows.size());
 	}
+	
+	@Test
+	public void testNonLatinDatabaseFilter() throws Exception {
+		String[] sql = {"create database 測試資料庫二"};
+		List<RowMap> rows = getRowsForDDLTransaction(sql, excludeDb("測試資料庫二"));
+		assertEquals(0, rows.size());
+	}
+	
 
 	@Test
 	public void testDatabaseChangeWithTableFilter() throws Exception {
@@ -468,5 +528,20 @@ public class DDLIntegrationTest extends MaxwellTestWithIsolatedServer {
 		assertEquals(2, rows.size());
 		assertTrue(rows.get(0).toJSON(ddlOutputConfig()).contains("\"type\":\"database-create\",\"database\":\"TestDatabaseCreate3\""));
 		assertTrue(rows.get(1).toJSON(ddlOutputConfig()).contains("\"type\":\"table-create\",\"database\":\"TestDatabaseCreate3\",\"table\":\"burger\""));
+	}
+	
+	@Test
+	public void testNonLatinDatabaseChangeWithTableFilter() throws Exception {
+		String[] sql = {
+				"create database 測試資料庫三",
+				"create table `測試資料庫三`.`表格一` ( 中文欄位 int )",
+				"create table `測試資料庫三`.`表格二` ( 中文欄位 int )"
+		};
+		List<RowMap> rows = getRowsForDDLTransaction(sql, excludeTable("表格一"));
+		assertEquals(2, rows.size());
+		assertTrue(rows.get(0).toJSON(ddlOutputConfig()).contains("\"type\":\"database-create\",\"database\":\"測試資料庫三\""));
+		assertTrue(rows.get(1).toJSON(ddlOutputConfig()).contains("\"type\":\"table-create\",\"database\":\"測試資料庫三\",\"table\":\"表格二\""));
+		// test if non-Latin column name outputs correctly
+		assertTrue(rows.get(1).toJSON(ddlOutputConfig()).contains("\"type\":\"int\",\"name\":\"中文欄位\""));
 	}
 }