+++ src/test/java/com/zendesk/maxwell/schema/ddl/DDLResolverTest.java
@@ -78,8 +78,8 @@ public class DDLResolverTest extends MaxwellTestWithIsolatedServer {
 		ResolvedTableCreate rc = c.resolve(getSchema());
 
 		assertThat(rc.def.charset, is("latin2"));
-		assertThat(((StringColumnDef) rc.def.getColumnList().get(0)).charset, is("latin2"));
-		assertThat(((StringColumnDef) rc.def.getColumnList().get(1)).charset, is("utf8"));
+		assertThat(((StringColumnDef) rc.def.getColumnList().get(0)).getCharset(), is("latin2"));
+		assertThat(((StringColumnDef) rc.def.getColumnList().get(1)).getCharset(), is("utf8"));
 
 	}
 
@@ -91,7 +91,7 @@ public class DDLResolverTest extends MaxwellTestWithIsolatedServer {
 		ResolvedTableCreate rc = c.resolve(getSchema());
 		assertThat(rc.def.getColumnList().size(), is(2));
 		assertThat(rc.def.getPKList().get(0), is("ii"));
-		assertThat(((StringColumnDef) rc.def.getColumnList().get(1)).charset, is("utf8"));
+		assertThat(((StringColumnDef) rc.def.getColumnList().get(1)).getCharset(), is("utf8"));
 	}
 
 	@Test